+++ src/Main.java
@@ -1,0 +1,10 @@
+import java.awt.*;
+public class Main {
+    public Main(){
+
+    }
+
+    public static void main(String[] args){
+
+    }
+}
