+++ src/Volvo240.java
@@ -1,7 +1,6 @@
 import java.awt.*;
 
 public class Volvo240 extends Car{
-
     private final static double trimFactor = 1.25;
 
     public Volvo240(){
@@ -12,22 +11,14 @@ public class Volvo240 extends Car{
         return enginePower * 0.01 * trimFactor;
     }
 
+    @Override
     public void incrementSpeed(double amount){
-	    currentSpeed = Math.min(getCurrentSpeed() + speedFactor() * amount,enginePower);
+	    currentSpeed = Math.min(getCurrentSpeed() + speedFactor() * amount, enginePower);
     }
 
-
+    @Override
     public void decrementSpeed(double amount){
         currentSpeed = Math.max(getCurrentSpeed() - speedFactor() * amount,0);
     }
 
-    // TODO fix this method according to lab pm
-    public void gas(double amount){
-        incrementSpeed(amount);
-    }
-
-    // TODO fix this method according to lab pm
-    public void brake(double amount){
-        decrementSpeed(amount);
-    }
 }
